requirements.txt
pandas
notebook
matplotlib
seaborn
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
types-PyYAML
scipy
Flask
Flask-Cors
opencv-python
scikit-learn
gdown
dvc
imutils
langchain==0.0.292 
openai==0.28.0 
datasets==2.10.1 
pinecone-client==2.2.4 
tiktoken==0.5.1
sentence_transformers
bitsandbytes
transformers @git+https://github.com/huggingface/transformers.git
peft @git+https://github.com/huggingface/peft.git
accelerate @git+https://github.com/huggingface/accelerate.git
py7zr
boto3
mypy-boto3-s3
fastapi==0.78.0
uvicorn==0.18.3
Jinja2==3.1.2
python-dotenv
wget
-e .